v0.9

+++ ewm-service/src/main/java/ru/practicum/api/privateAPI/comment/PrivateCommentServiceImpl.java
@@ -60,12 +60,9 @@ public class PrivateCommentServiceImpl implements PrivateCommentService {
         if (comment.getAuthor() == null) {
             throw new NotFoundException("Comment does not have author");
         }
-       // if (!comment.getAuthor().getId().equals(userId)) {
-
-       /*if (comment.getAuthor().getId() != userId) {
-
-            throw new ForbiddenException("Only author can delete comment");
-        }*/
+        if (!comment.getAuthor().getId().equals(userId)) {
+           throw new ForbiddenException("Only author can delete comment");
+        }
         commentRep.deleteById(comId);
     }
